perf(web): finish lazy-modal split — eager imports + chunk cycle fix
Verifier surfaced two reasons the Tiptap chunk was still being downloaded
at /app boot despite the lazy wrappers landed in the previous commit:

1. `Sidebar` (rendered on every /app load), `routes/app/tasks.tsx`, and
   `routes/app/calendar.tsx` were still importing the eager `task-modal` /
   `add-task-modal` modules. Switched to the `.lazy` wrappers.

2. The Vite manualChunks rule for `react-core` used a naive
   `id.includes("/react/")` match, which silently captured `@tiptap/react`,
   `@radix-ui/react-*`, and `@tanstack/react-router` modules. Pulling those
   into react-core inverted the dependency graph: a chunk that's loaded
   eagerly on first paint depended on the lazy chunks, forcing Rollup to
   emit `import "./tiptap..."` side-effect imports into entry chunks via
   `hoistTransitiveImports`. Tightened the regex to require
   `/node_modules/react/` etc., and disabled `hoistTransitiveImports` so
   transitive dynamic deps stay on-demand.

Net result: tiptap (~417 kB) is no longer fetched on /app boot. It loads
when the user opens an Add Task or Task modal, prefetched on idle. The
react-core chunk also shrank from 237 kB → 187 kB now that it's actually
just react/react-dom/scheduler.

The "Circular chunk" warnings that were emitted on every build are also
gone now that the manual-chunks regex is correct.
