+++ apps/web/src/routes/app/tasks.tsx
@@ -4,8 +4,8 @@ import { format, isToday, isTomorrow, isPast, parseISO } from "date-fns";
 import type { Task } from "@open-sunsama/types";
 import { useInfiniteSearchTasks } from "@/hooks/useInfiniteSearchTasks";
 import { useCompleteTask, useReorderTasks } from "@/hooks/useTasks";
-import { TaskModal } from "@/components/kanban/task-modal";
-import { AddTaskModal } from "@/components/kanban/add-task-modal";
+import { TaskModal } from "@/components/kanban/task-modal.lazy";
+import { AddTaskModal } from "@/components/kanban/add-task-modal.lazy";
 import { TaskGroup } from "@/components/tasks/task-group";
 import { TaskShortcutsHandler } from "@/components/task-shortcuts-handler";
 import { Button } from "@/components/ui";
